+++ src/library.php
@@ -6,17 +6,16 @@ namespace XPSPL;
  * that can be found in the LICENSE file.
  */
 
-use processor\signal\Module_Load_Failure,
-    processor\event\Error;
+use XPSPL\exception\Module_Load_Failure;
 
 /**
  * Library
  * 
  * Loads and tracks XPSPL modules.
  */
-class Library {
+class Library extends Database {
 
-    use Storage, Singleton;
+    use Singleton;
 
     /**
      * Loads a XPSPL module.
@@ -24,37 +23,19 @@ class Library {
      * @param  string  $name  Module name.
      * @param  string|null  $dir  Location of the module. 
      * 
-     * @return  boolean
+     * @return  void
      */
-    public function load($name, $dir = null) 
+    public function load($name, $dir = XPSPL_MODULE_DIR) 
     {
         // already loaded
-        if (isset($this->_storage[$name])) return true;
-        if ($dir === null) {
-            $dir = XPSPL_MODULE_DIR;
-        } else {
-            if (!is_dir($dir)) {
-                emit(new Module_Load_Failure(),  new Error(sprintf(
-                    "Module directory %s does not exist", $dir
-                )));
-            }
+        if (isset($this->_storage[$name])) return;
+        if (!is_dir($dir)) {
+            throw new Module_Load_Failure(sprintf(
+                "Module directory %s does not exist", $dir
+            ));
         }
-        if (is_dir($dir.'/'.$name)) {
-            $path = $dir.'/'.$name;
-            if (file_exists($path.'/__autoload.php')) {
-                // keep history of what has been loaded
-                $this->_storage[$name] = true;
-                require $path.'/__autoload.php';
-            } else {
-                emit(new Module_Load_Failure(), new Error(
-                    "Module does not have an __autoload file"
-                ));
-            }
-        } else {
-            emit(new Module_Load_Failure(), new Error(sprintf(
-                "Module %s does not exist", $name
-            )));
-        }
-        return true;
+        $path = $dir.'/'.$name;
+        $this->_storage[$name] = true;
+        require_once $path.'/__init__.php';
     }
 }